bot/requirements.txt
pydle
duckduckgo3
asyncio